backend/requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyJWT==2.9.0
redis==5.2.0
sqlparse==0.5.1
typing_extensions==4.12.2
wheel==0.45.0
